+++ bin/store/src/organizations/organization_controller.rs
@@ -1,7 +1,7 @@
 use actix_web::{web, HttpRequest, HttpResponse, Responder};
 use serde::{Deserialize, Serialize};
 
-use crate::auth::structs::Session;
+// use crate::auth::structs::Session;
 use crate::organizations::auth_service::{auth, root_auth};
 use crate::organizations::organization_service::register;
 use crate::organizations::structs::Register;
@@ -23,9 +23,9 @@ pub struct AuthData {
 pub struct OrganizationsController;
 
 impl OrganizationsController {
-    pub fn new() -> Self {
-        Self
-    }
+    // pub fn new() -> Self {
+    //     Self
+    // }
 
     pub async fn register(data: web::Json<Register>) -> impl Responder {
         if data.0.account_id.is_empty() || data.0.account_secret.is_empty() {